check users validity before service

+++ backend/controllers/bank_account.controller.js
@@ -2,6 +2,9 @@ const BankAccount = require('../models/bank_account.model');
 
 exports.createBankAccount = async (req, res) => {
   try {
+    const userId = req.userId; 
+    req.body.user_id = userId; 
+    console.log(userId);
     const bankAccount = await BankAccount.create(req.body);
     res.status(201).json(bankAccount);
   } catch (err) {
@@ -12,7 +15,9 @@ exports.createBankAccount = async (req, res) => {
 
 exports.getAllBankAccounts = async (req, res) => {
   try {
-    const bankAccounts = await BankAccount.findAll();
+    const userId = req.userId; 
+    console.log(userId);
+    const bankAccounts = await BankAccount.findAll({ where: { user_id: userId } });
     res.json(bankAccounts);
   } catch (err) {
     console.error('Error getting bank accounts:', err);
@@ -23,9 +28,10 @@ exports.getAllBankAccounts = async (req, res) => {
 exports.getBankAccountById = async (req, res) => {
   const bankAccountId = req.params.id;
   try {
-    const bankAccount = await BankAccount.findByPk(bankAccountId);
+    const userId = req.userId; 
+    const bankAccount = await Expense.findOne({ where: { id: bankAccountId, user_id: userId } });
     if (!bankAccount) {
-      return res.status(404).json({ message: 'Bank account not found' });
+      return res.status(404).json({ message: 'Bank Account not found or does not belong to the authenticated user' });
     }
     res.json(bankAccount);
   } catch (err) {
@@ -37,6 +43,11 @@ exports.getBankAccountById = async (req, res) => {
 exports.updateBankAccount = async (req, res) => {
   const bankAccountId = req.params.id;
   try {
+    const userId = req.userId; 
+    const bankAccount = await Expense.findOne({ where: { id: bankAccountId, user_id: userId } });
+    if (!bankAccount) {
+      return res.status(404).json({ message: 'Bank Account not found or does not belong to the authenticated user' });
+    }
     const [updatedRowsCount, updatedRows] = await BankAccount.update(req.body, {
       where: { id: bankAccountId },
       returning: true,
@@ -54,6 +65,11 @@ exports.updateBankAccount = async (req, res) => {
 exports.deleteBankAccount = async (req, res) => {
   const bankAccountId = req.params.id;
   try {
+    const userId = req.userId; 
+    const bankAccount = await Expense.findOne({ where: { id: bankAccountId, user_id: userId } });
+    if (!bankAccount) {
+      return res.status(404).json({ message: 'Bank Account not found or does not belong to the authenticated user' });
+    }
     const deletedRowsCount = await BankAccount.destroy({ where: { id: bankAccountId } });
     if (deletedRowsCount === 0) {
       return res.status(404).json({ message: 'Bank account not found' });
